python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hellojohn"
version = "1.0.0"
description = "HelloJohn authentication SDK for Python (FastAPI, Flask)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["authentication", "jwt", "fastapi", "flask", "hellojohn"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Framework :: FastAPI",
  "Framework :: Flask",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "PyJWT>=2.8.0",
  "cryptography>=42.0.0",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.110.0"]
flask = ["flask>=3.0.0"]
all = ["fastapi>=0.110.0", "flask>=3.0.0"]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "fastapi[all]>=0.110.0",
  "flask>=3.0.0",
  "respx>=0.21.0"
]

[tool.hatch.build.targets.wheel]
packages = ["hellojohn"]

